Handle upstream join streaming and simplify some logic (#405)
## Summary

Per the conversation on
[Slack](https://zipline-2kh4520.slack.com/archives/C0880ECQ0EN/p1739982805433429).
We don't currently have an example where a `GroupBy` takes in both a
streaming and a batch source (mostly are a single source, and a few that
are not are batch only), but the `.some()`/any logic should be
[right](https://zipline-2kh4520.slack.com/archives/C0880ECQ0EN/p1739830806387109)

Conf | Before | After
--- | --- | ---
Join |
![image](https://github.com/user-attachments/assets/0f989d85-f9a5-4612-9355-1c4f8efd5d7d)
|
![image](https://github.com/user-attachments/assets/dc0d9613-28b4-4156-921e-24210b52d6d3)
GroupBy (no change) |
![image](https://github.com/user-attachments/assets/eb2727e3-a3ed-4ef5-8faf-3b594f7dc9d5)
|
![image](https://github.com/user-attachments/assets/957812d4-da96-42fe-b361-7804476f416c)


## Checklist
- [ ] Added Unit Tests
- [ ] Covered by existing CI
- [ ] Integration tested
- [ ] Documentation update


<!-- av pr metadata
This information is embedded by the av CLI when creating PRs to track
the status of stacks when using Aviator. Please do not delete or edit
this section of the PR.
```
{"parent":"main","parentHead":"","trunk":"main"}
```
-->


---
- To see the specific tasks where the Asana app for GitHub is being
used, see below:
  - https://app.asana.com/0/0/1209446053233127

<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

- **Refactor**
- Improved the application’s handling of streaming content for enhanced
reliability and performance. The updates streamline the evaluation of
live data and combined streams, contributing to a more consistent and
responsive user experience.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->

Co-authored-by: Sean Lynch <techniq35@gmail.com>