Fix sort when client-side routed and simplify (#414)
## Summary

Not sure the root cause of using `queryParameters()` within the
drift/distributions route is not reactive when client-side routed
(clicking on the Observability tab or clicking between Drift and
Distributions) but works if you refresh the page. It is reactive within
`ActionsButtons` and generally everywhere else.

At some point I'll workup a minimal repo and report on
https://github.com/paoloricciuti/sveltekit-search-params, but this
approach is both simpler and resolves the issue (and still uses the same
param config).

My initial guess is it might be due to subscribing to the same params
with 2 different instances on the same page, but not positive without
investigating (and possibly not related).

[Asana issue](https://app.asana.com/0/1208932365725930/1209459569820188)


https://github.com/user-attachments/assets/3b825bae-484b-4dba-8c35-0e84b0b65ce0

## Checklist
- [ ] Added Unit Tests
- [ ] Covered by existing CI
- [ ] Integration tested
- [ ] Documentation update


<!-- av pr metadata
This information is embedded by the av CLI when creating PRs to track
the status of stacks when using Aviator. Please do not delete or edit
this section of the PR.
```
{"parent":"main","parentHead":"","trunk":"main"}
```
-->


---
- To see the specific tasks where the Asana app for GitHub is being
used, see below:
  - https://app.asana.com/0/0/1209459569820188

Co-authored-by: Sean Lynch <techniq35@gmail.com>