Update LayerChart, LayerStack, and Svelte UX packages to 1.0.0 (#440)
## Summary

Just released 1.0 releases for LayerChart, LayerStack, and Svelte UX
packages which have Svelte 3-5 and Tailwind 3 compatibility
([announcement](https://bsky.app/profile/techniq.dev/post/3lj3sjabpns2z))
(LayerChart 1.0 is also 98% Tailwind 4
[compatibility](https://github.com/techniq/layerchart/issues/388)).

This mostly paves the way to start on `next` branches / `2.0` releases
which will leverage Svelte 5 runes/snippets and Tailwind 4 (along with a
lot of other planned features).

## Checklist
- [ ] Added Unit Tests
- [ ] Covered by existing CI
- [ ] Integration tested
- [ ] Documentation update


<!-- av pr metadata
This information is embedded by the av CLI when creating PRs to track
the status of stacks when using Aviator. Please do not delete or edit
this section of the PR.
```
{"parent":"main","parentHead":"","trunk":"main"}
```
-->

Co-authored-by: Sean Lynch <techniq35@gmail.com>