Update svelte-inspect-value and use new css/theme variables (#453)
## Summary

Using the new
[recently](https://github.com/ampled/svelte-inspect-value/issues/23#issuecomment-2690391648)
released extended css/theme variables which allows to mute the hierarchy
lines but not the caret.

This also affected some of the default colors, so I added more theme
colors to (mostly) align to the previous colors (without these the
values were mostly white).

Before | After
--- | ---

![image](https://github.com/user-attachments/assets/c012db55-5cff-4b02-a719-08298d7b0783)
|
![image](https://github.com/user-attachments/assets/52281311-4c69-449c-91e4-2aa67dbc0bf0)


## Checklist
- [ ] Added Unit Tests
- [ ] Covered by existing CI
- [ ] Integration tested
- [ ] Documentation update


<!-- av pr metadata
This information is embedded by the av CLI when creating PRs to track
the status of stacks when using Aviator. Please do not delete or edit
this section of the PR.
```
{"parent":"main","parentHead":"","trunk":"main"}
```
-->

Co-authored-by: Sean Lynch <techniq35@gmail.com>