Stop using cloud_gcp_submitter jar and just use cloud_gcp jar, and add `--gcp` for `fetch` (#466)
## Summary
^^^

Removing usage and references to the cloud_gcp_submitter jar as we don't
need it. And added an additional `--gcp` arg so that it can be used in
the `fetch` command to set which cloud provider's implementations we
should use. Ex: `fetch` + `--gcp` should be pulling from the `BigTable`
kv store code

## Checklist
- [ ] Added Unit Tests
- [ ] Covered by existing CI
- [ ] Integration tested
- [ ] Documentation update



<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

- **New Features**
- Added a new command-line option for specifying GCP settings, allowing
users to easily switch modes.
- Introduced an automated script for creating Python wheel packages
during artifact builds.

- **Refactor**
- Updated module imports and renamed constants for improved clarity and
consistency.
- Simplified logic for handling jar file paths and streamlined the
artifact build process.
  - Removed obsolete build configurations and redundant commands.

- **Documentation**
- Enhanced developer notes with improved formatting for better
readability.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->