fix: bootstrap unit tests should be deterministic (#487)
## Summary

- In certain tests we will generate some dummy data with random values.
We proceed to use this data to do correctness checks. However, since
dataframes are lazily evaluated, whenever we reference the value of the
dataframe, the random generation kicks in. Thus, we sometimes get
incorrect unit test results based on the data that gets generated at the
time of referencing the dataframe. The fix is to first save the
dataframe to a table, and load it back immediately to preserve the
values that were generated the first time.

## Checklist
- [ ] Added Unit Tests
- [ ] Covered by existing CI
- [ ] Integration tested
- [ ] Documentation update
<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **Refactor**
- Standardized the approach to loading table data in test cases for
improved consistency and maintainability.
- Updated variable names to enhance clarity regarding the data being
handled.
  
- **Tests**
- Improved test logic with clearer variable naming and a refined process
to ensure tests reflect the most recent data state after saving.
  - Minor formatting adjustment in the test case for better readability.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->

<!-- av pr metadata
This information is embedded by the av CLI when creating PRs to track
the status of stacks when using Aviator. Please do not delete or edit
this section of the PR.
```
{"parent":"main","parentHead":"","trunk":"main"}
```
-->

---------

Co-authored-by: Thomas Chow <thomaschow369@gmail.com>