Canary Automated Testing  (#471)
## Summary

This updates the Push To Canary workflow to push the updated jars and
wheel to the cloud canaries and run integration tests.

The jobs are as follows:
* **build_artifacts** 
builds the jars and wheel from the main branch and uploads them to
github for use in separate jobs simultaneously
* **push_to_aws**
downloads the jars and wheel from github and uploads them to the s3
zipline-artifacts-canary bucket under version "candidate"
* **push_to_gcp**
downloads the jars and wheel from github and uploads them to the gcs
zipline-artifacts-canary bucket under version "candidate"
* **run_aws_integration_tests**
runs the quickstart compile and backfill based on
[scripts/distribution/run_aws_quickstart.sh](https://github.com/zipline-ai/chronon/blob/main/scripts/distribution/run_aws_quickstart.sh)
* **run_gcp_integration_tests**
runs the quickstart compile, backfill, group-by-upload, upload-to-kv,
metadata-upload, and fetch based on
[scripts/distribution/run_gcp_quickstart.sh](https://github.com/zipline-ai/chronon/blob/main/scripts/distribution/run_gcp_quickstart.sh)
* **clean_up_artifacts**
removes the jars and wheel from github to minimize the storage costs.
* **merge_to_main-passing-tests**
if the tests all pass, this will merge the latest commits to the branch
"main-passing-tests". This also creates a txt file artifact marking
which commit sha passed which is saved by github.
* **on_fail_notify_slack**
if the tests fail, this will send a notification to the
"integration-tests" channel of our Slack.

A test run of the workflow can be seen here:
https://github.com/zipline-ai/chronon/actions/runs/13913346353

## Checklist
- [ ] Added Unit Tests
- [ ] Covered by existing CI
- [x] Integration tested
- [ ] Documentation update



<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

## Summary by CodeRabbit

- **Chores**
- Enhanced our release pipeline to streamline package generation and
distribution.
	- Automated deployment of new build packages across AWS and GCP.
	- Introduced cleanup routines to optimize the post-deployment process.
- Updated job structure for improved artifact management and deployment.
- Added new jobs for building artifacts and uploading them to cloud
storage, including `build_artifacts`, `push_to_aws`, `push_to_gcp`, and
`clean_up_artifacts`.
	- Defined outputs for generated artifacts to improve tracking.
- Improved flexibility in deployment scripts for different environments
(dev and canary).
- Updated scripts to conditionally handle resource management based on
environment settings.
- Introduced environment-specific configurations for AWS and GCP
operations.
	- Added integration testing jobs for AWS and GCP post-deployment.
- Enhanced scripts for AWS and GCP to support environment-specific
operations and configurations.
- Introduced a mechanism to switch between different environments
affecting resource management and command configurations.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->