Finish Canary Runs (#564)
## Summary
This change removes cancel-in-progress so that the concurrency rules
will wait for any existing jobs to complete before starting a new run of
the workflow. [Concurrency
Rules](https://docs.github.com/en/enterprise-cloud@latest/actions/writing-workflows/choosing-what-your-workflow-does/control-the-concurrency-of-workflows-and-jobs#using-concurrency-in-different-scenarios)
state "This means that there can be at most one running and one pending
job in a concurrency group at any time. When a concurrent job or
workflow is queued, if another job or workflow using the same
concurrency group in the repository is in progress, the queued job or
workflow will be pending."

This change also makes the commit sha a link to the workflow run logs.


## Checklist
- [ ] Added Unit Tests
- [ ] Covered by existing CI
- [ ] Integration tested
- [ ] Documentation update



<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

## Summary by CodeRabbit

- **Chores**
- Enhanced our update process to support concurrent operations, allowing
new updates to proceed alongside ongoing ones without interruption.
- Improved clarity of failure notifications sent to Slack by formatting
the commit SHA as a clickable link to the specific run in GitHub
Actions.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->