Fix mode configs logic for when no modeConfigs are set (#613)
## Summary

Noticed that when I ran compile on Etsy's conf, their conf looked like
this:

```
   conf=ConfigProperties(
        common={
...        },
        )
```

With the current code, this ended up with our Submitter not setting any
job confs at all. This is because we were checking for the `modeConfigs`
set, and if not we'd do nothing. This is a bug. Fixing that logic

## Checklist
- [ ] Added Unit Tests
- [ ] Covered by existing CI
- [ ] Integration tested
- [ ] Documentation update



<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

## Summary by CodeRabbit

- **Refactor**
- Streamlined the internal process for retrieving mode configurations,
reducing its complexity and improving efficiency while maintaining
existing functionality.
- **New Features**
- Introduced new JSON configuration files for the
`quickstart.purchases.v1` dataset, enhancing data management and
execution settings.
- Added a new test class for validating the functionality of the
`JobSubmitter` class, ensuring robustness in configuration parsing and
argument retrieval.
- **Chores**
- Updated test suite configurations to improve testing structure and
accessibility.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->