feat: GCPFormatProvider should do the default checks first  (#617)
## Summary

- Now that we use Polaris for etsy, there's a case where we don't go to
BigQuery to find a table. We'll need to bring in the normal iceberg
check to do that.

## Checklist
- [ ] Added Unit Tests
- [ ] Covered by existing CI
- [ ] Integration tested
- [ ] Documentation update
<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **Refactor**
- Enhanced the cloud integration’s data formatting process by
streamlining the resolution strategy, checking for pre-existing formats
before applying a fallback.
- Improved accessibility of the `sparkSession` parameter in the
`DefaultFormatProvider` class for direct access from instances.
- Improved overall consistency and efficiency in data handling for
cloud-based operations.
- Updated the instantiation method for the `GcpFormatProvider` class in
tests for clarity.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->

<!-- av pr metadata
This information is embedded by the av CLI when creating PRs to track
the status of stacks when using Aviator. Please do not delete or edit
this section of the PR.
```
{"parent":"main","parentHead":"","trunk":"main"}
```
-->

---------

Co-authored-by: Thomas Chow <thomaschow369@gmail.com>